requirements.txt
dj-database-url==0.5.0
Django==2.2.4
gunicorn==19.9.0
psycopg2==2.7.5
pytz==2018.5
whitenoise==4.0
django-taggit==0.22.2
Django-crispy-forms==1.7.2

certifi==2019.9.11
chardet==3.0.4
django-mathfilters==0.4.0
djangorestframework==3.10.3
idna==2.8
python-memcached==1.59
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6

